feat: adjust base types for global usage
